# backend/api/openapi/swagger.yaml
definitions:
...
  sico-backend_internal_transport_http_dto_common.Attachment:
    properties:
      id:
        type: integer
      name:
        type: string
      sasUrl:
        type: string
      size:
        type: integer
      type:
        type: string
      uri:
        type: string
    type: object
...
  sico-backend_internal_transport_http_dto_scheduledtask.CreateScheduledTaskRequest:
    properties:
      agentInstanceId:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_common.Attachment'
        type: array
      cronExpression:
        type: string
      enabled:
        type: boolean
      extraInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTaskExtraInfo'
      message:
        type: string
      name:
        type: string
      timezone:
        type: string
    required:
    - agentInstanceId
    - cronExpression
    - message
    - name
    - timezone
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.CreateScheduledTaskResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTask'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.DeleteScheduledTaskRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.DeleteScheduledTaskResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.GetScheduledTaskResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTask'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.ListScheduledTasksData:
    properties:
      hasNext:
        type: boolean
      tasks:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTask'
        type: array
      total:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.ListScheduledTasksResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ListScheduledTasksData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTask:
    properties:
      agentInstanceId:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_common.Attachment'
        type: array
      createdAt:
        type: integer
      creatorUsername:
        type: string
      cronExpression:
        type: string
      enabled:
        type: boolean
      extraInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTaskExtraInfo'
      id:
        type: integer
      lastRunAt:
        type: integer
      message:
        type: string
      name:
        type: string
      nextRunAt:
        type: integer
      timezone:
        type: string
      updatedAt:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTaskExtraInfo:
    properties:
      sendEmailOnComplete:
        type: boolean
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.UpdateScheduledTaskRequest:
    properties:
      agentInstanceId:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_common.Attachment'
        type: array
      cronExpression:
        type: string
      enabled:
        type: boolean
      extraInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTaskExtraInfo'
      id:
        type: integer
      message:
        type: string
      name:
        type: string
      timezone:
        type: string
    required:
    - agentInstanceId
    - cronExpression
    - id
    - message
    - name
    - timezone
    type: object
  sico-backend_internal_transport_http_dto_scheduledtask.UpdateScheduledTaskResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ScheduledTask'
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Sico AI Agent Platform API
  title: Sico API
  version: "1.0"
paths:
...
  /api/sico/scheduled-tasks:
    delete:
      parameters:
      - description: Delete Scheduled Task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.DeleteScheduledTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.DeleteScheduledTaskResponse'
      security:
      - BearerAuth: []
      tags:
      - scheduled-tasks
    get:
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.GetScheduledTaskResponse'
      security:
      - BearerAuth: []
      tags:
      - scheduled-tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Scheduled Task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.CreateScheduledTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.CreateScheduledTaskResponse'
      security:
      - BearerAuth: []
      tags:
      - scheduled-tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Scheduled Task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.UpdateScheduledTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.UpdateScheduledTaskResponse'
      security:
      - BearerAuth: []
      tags:
      - scheduled-tasks
  /api/sico/scheduled-tasks/list:
    get:
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_scheduledtask.ListScheduledTasksResponse'
      security:
      - BearerAuth: []
      tags:
      - scheduled-tasks
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"